rent_themes/requirements.txt
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
black==22.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0
django-cors-headers==4.1.0
djangorestframework==3.13.1
drf-yasg==1.21.5
executing==0.8.2
idna==3.4
inflection==0.5.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==23.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pytz==2021.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.0
traitlets==5.1.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.5
